Now Pabito is excluded from censorship

+++ bot.py
@@ -225,13 +225,16 @@ async def on_message(message):
         await tree.sync()
         await message.channel.send("Zsynchronizowano drzewo!")
 
-    if time[0] < 20 and time[0] >= 5 and message.author.id != 421932366802714625:
+    if (time[0] < 20 and time[0] >= 5 and message.author.id != 421932366802714625) or message.author.id == 771078648140791808:
         for i in bannedWords:
             if msgLowercaseNoPolish.find(i) != -1:
                 remove = True
 
-    if remove:
-        toRemove = await message.channel.send("<@"+str(sender.id)+">!!! Zgodnie z paragrafem §1.8 na kanale <#935612476156936272> o godzinie "+timeNow+" czasu polskiego panuje bezwzględny zakaz używania przekleństw (z wyjątkami opisanymi w tym podpunkcie oraz za wyjątkiem boskiego Pabito). W związku z powyższym wiadomość została usunięta. Pilnuj się!")
+    if remove and message.author.id != bot.user.id:
+        if message.author.id == 771078648140791808:
+            toRemove = await message.channel.send("<@"+str(sender.id)+">!!! Ty dziwko jebana! Nie wolno ci przekilać, bo Pablo ci zabronił, a jego słowo jest święte (czy coś), w związku z tym wiadomość została usunięta. Pilnuj się!")
+        else:
+            toRemove = await message.channel.send("<@"+str(sender.id)+">!!! Zgodnie z paragrafem §1.8 na kanale <#935612476156936272> o godzinie "+timeNow+" czasu polskiego panuje bezwzględny zakaz używania przekleństw (z wyjątkami opisanymi w tym podpunkcie oraz za wyjątkiem boskiego Pabito). W związku z powyższym wiadomość została usunięta. Pilnuj się!")
         await message.delete()
         await toRemove.delete(delay=15)
     elif message.channel.id == env.COUNTING_CHANNEL and message.author.id != bot.user.id:
